Сделал прием новой модели, исправил отображение данных новой модели

+++ HxModel/Models/FCSState.cs
@@ -6,7 +6,7 @@ namespace HxModel.Models
 	[StructLayout(LayoutKind.Sequential,Pack = 1)]
 	struct FCSState
 	{
-		uint Mode;  // режим работы системы упрвления: 0 - Direct, 1 - AugmentedStability, 2 - TrimmeredStick(Classic), 3 - Auto
+		double Mode;  // режим работы системы упрвления: 0 - Direct, 1 - AugmentedStability, 2 - TrimmeredStick(Classic), 3 - Auto
 
 		// уставочные параметры системы управления, обрабатываются только при Mode = Auto :
 		FCSValue HorSpeedReq; // заданная горизонтальная скорость (приборная), м/с
